requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
dnspython==2.4.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
huggingface-hub==0.16.4
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
multidict==6.0.4
nltk==3.8.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
priority==2.0.0
pycparser==2.21
pydantic==2.0.2
pydantic-settings==2.0.1
pydantic_core==2.1.2
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyVimeo==1.1.0
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tinydb==4.8.0
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.65.0
transformers==4.31.0
tuspy==1.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wsproto==1.2.0
yarl==1.9.2
torch==2.0.1